refactor Android timeout build option

+++ android/src/main/java/com/sourcepoint/reactnativecmp/arguments/Arguments.kt
@@ -161,9 +161,9 @@ fun WritableMap.putArray(name: String, value: Iterable<*>) {
   })
 }
 
-fun ReadableMap.getLongOrNull(name: String) =
+fun ReadableMap.getDoubleOrNull(name: String) =
   if (hasKey(name) && !isNull(name)) {
-    getLong(name)
+    getDouble(name)
   } else {
     null
   }
